fix(fip0118): address reviewer feedback on solstice CLI and itests
- Extract packUpgradeQualityMessages + sectorLoc to package level for
  testability; add sectors_upgrade_quality_test.go covering grouping,
  splitting, and CBOR round-trip
- Add --max-sectors negative check
- Resolve eth block tipset from ethBlockA.Hash to eliminate reorg race
- Add TestMigrationNV29SolsticeFullQaHelperSnapExtend to verify
  SectorIsFullQaPower against on-chain QA power across snap/extend; assert Expiration > PowerBaseEpoch invariant
- Split solstice_fault_fees_test.go into three focused files
